+++ xmlschema/validators/xsdbase.py
@@ -19,8 +19,7 @@ from ..names import XSD_ANNOTATION, XSD_APPINFO, XSD_DOCUMENTATION, XML_LANG, \
     XSD_OVERRIDE, XSD_NOTATION_TYPE, XSD_DECIMAL
 from ..etree import is_etree_element, etree_tostring
 from ..helpers import get_qname, local_name, get_prefixed_qname
-from ..dataobjects import DataElement
-from ..converters import DataElementConverter
+from .. import dataobjects
 from .exceptions import XMLSchemaParseError, XMLSchemaValidationError
 
 XSD_TYPE_DERIVATIONS = {'extension', 'restriction'}
@@ -769,14 +768,14 @@ class ValidationMixin:
         for an attribute or a simple type validators, or an ElementTree's Element otherwise.
         :param use_defaults: indicates whether to use default values for filling missing data.
         :param namespaces: is an optional mapping from namespace prefix to URI.
-        :raises: :exc:`XMLSchemaValidationError` if XML *data* instance is not a valid.
+        :raises: :exc:`XMLSchemaValidationError` if XML *data* instance is not valid.
         """
         for error in self.iter_errors(source, use_defaults=use_defaults, namespaces=namespaces):
             raise error
 
     def is_valid(self, source, use_defaults=True, namespaces=None):
         """
-        Like :meth:`validate` except that do not raises an exception but returns ``True`` if
+        Like :meth:`validate` except that does not raise an exception but returns ``True`` if
         the XML document is valid, ``False`` if it's invalid.
 
         :param source: the source of XML data. For a schema can be a path \
@@ -838,6 +837,24 @@ class ValidationMixin:
 
         return (result, errors) if validation == 'lax' else result
 
+    def to_objects(self, source, with_bindings=False, **kwargs):
+        """
+        Decodes XML data to Python data objects.
+
+        :param source: the XML data. Can be a string for an attribute or for a simple \
+        type components or a dictionary for an attribute group or an ElementTree's \
+        Element for other components.
+        :param with_bindings: if `True` is provided the decoding is done using \
+        :class:`DataBindingConverter` that used XML data binding classes. For \
+        default the objects are intances of :class:`DataElement` and uses the \
+        :class:`DataElementConverter`.
+        :param kwargs: other optional keyword arguments for the method \
+        :func:`iter_decode`, except the argument *converter*.
+        """
+        if with_bindings:
+            return self.decode(source, converter=dataobjects.DataBindingConverter, **kwargs)
+        return self.decode(source, converter=dataobjects.DataElementConverter, **kwargs)
+
     def encode(self, obj, validation='strict', **kwargs):
         """
         Encodes data to XML.
@@ -853,9 +870,6 @@ class ValidationMixin:
         component, or also if it's invalid when ``validation='strict'`` is provided.
         """
         check_validation_mode(validation)
-        if 'converter' not in kwargs and isinstance(obj, DataElement):
-            kwargs['converter'] = DataElementConverter
-
         result, errors = None, []
         for result in self.iter_encode(obj, validation=validation, **kwargs):  # pragma: no cover
             if not isinstance(result, XMLSchemaValidationError):
@@ -881,7 +895,7 @@ class ValidationMixin:
 
     def iter_encode(self, obj, validation='lax', **kwargs):
         """
-        Creates an iterator for Encode data to an Element.
+        Creates an iterator for encoding data to an Element tree.
 
         :param obj: The data that has to be encoded.
         :param validation: The validation mode. Can be 'lax', 'strict' or 'skip'.